Add check for file X window env lock in order to allow removal

+++ start.sh
@@ -1,5 +1,8 @@
 # Clean up any unreleased virtual frame buffer instances and prepare the new one
-rm /tmp/.X9-lock
+
+if [ -f /tmp/.X9-lock ]; then
+  rm /tmp/.X9-lock
+fi
 
 # Create new virtual frame buffer with X-Window environment
 Xvfb -ac -screen scrn 1280x960x24 :9.0 &
